Move internal/csr to pkg/csr
