+++ issuers/cfssl/cfssl.go
@@ -13,7 +13,7 @@ import (
 	"github.com/cloudflare/cfssl/signer"
 
 	"github.com/johanbrandhorst/certify"
-	"github.com/johanbrandhorst/certify/internal/csr"
+	"github.com/johanbrandhorst/certify/ext/csr"
 )
 
 // Issuer implements the Issuer interface
